🛠️ Refactor: model imports and add MobileNetV1 and MobileNetV2 implementations
- Updated __init__.py to import MobileNetV1 and MobileNetV2 from new mobilenets module.
- Created mobilenets.py with implementations for MobileNetV1 and MobileNetV2.
- Created resnets.py with ResNet implementation and associated methods.
- Modified training module to reflect a more specific training message for MobileNetV1 on CIFAR-10.
